jest initial server set up
