+++ server/config/db.js
@@ -1,9 +1,11 @@
 const mongoose = require('mongoose');
+
+let conn;
 const connectDB = async () => {
   //Mongoose 6 default: useNewUrlParser, useUnifiedTopology, and useCreateIndex are true, and useFindAndModify is false
-  const conn = await mongoose.connect(process.env.MONGO_URI, {});
+  conn = await mongoose.connect(process.env.MONGO_URI, {});
 
   console.log(`MongoDB Connected: ${conn.connection.host}`);
 };
 
-module.exports = connectDB;
+module.exports = { connectDB, conn };
