+++ server/package.json
@@ -6,8 +6,12 @@
   "engines": {
     "node": ">=16.14.0"
   },
+  "jest": {
+    "testEnvironment": "node"
+  },
   "scripts": {
     "start": "node server",
+    "test": "jest --watchAll --detectOpenHandles",
     "dev": "nodemon"
   },
   "repository": {
@@ -21,8 +25,14 @@
   },
   "homepage": "https://github.com/chamoo334/fcfs-api#readme",
   "devDependencies": {
+    "@babel/core": "^7.17.9",
+    "@babel/node": "^7.16.8",
+    "@babel/plugin-transform-runtime": "^7.17.0",
+    "@babel/preset-env": "^7.16.11",
     "concurrently": "^7.0.0",
-    "nodemon": "^2.0.15"
+    "jest": "^27.5.1",
+    "nodemon": "^2.0.15",
+    "supertest": "^6.2.2"
   },
   "dependencies": {
     "@aws-sdk/client-ses": "^3.58.0",
